requirements.txt
fastapi
pydantic
pyjwt
uvicorn
python-jose
sqlalchemy
pymysql
python-dotenv
mysql-connector-python
passlib
pydantic[email]
bcrypt